+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-20T13:21:07.678Z",
+    "2026-05-20T17:38:29.061Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {