+++ packages/orq-rc/RELEASES.md
@@ -9359,3 +9359,13 @@ Based on:
 - [typescript v4.10.0-rc.22] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.22 - packages/orq-rc
+
+## 2026-05-20 17:41:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.4 (2.884.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.23] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.23 - packages/orq-rc